Create Construction Function
In this example, MyContract is a simple smart contract that has a single state variable called myVariable. The constructor function takes one parameter, initialValue, which is used to set the initial value of myVariable when the contract is deployed.

The updateVariable function can be called by anyone to update the value of myVariable. Note that because myVariable is marked as public, Solidity automatically generates a getter function that can be used to retrieve its value from outside the contract.