Local Variables
Local variables are declared inside a function and are only accessible within that function. They are used to store temporary values that are needed for computation.

In this case, 

uint256 result = 2 + 2

"result" is a uint variable that takes 2 + 2 as a value.  

Returning this variable will always result in the number "4".