Modifier Functions
A modifier is a special function that can be used to modify the behavior of other functions in a contract. A modifier function can be used to add additional checks or requirements before executing a function or to modify the state of the contract before or after executing a function.

In this example, the onlyOwner modifier function is used to restrict access to the doSomething function. The modifier checks whether the caller of the function is the same as the owner of the contract. If the caller is not the owner, the function call will fail with an error message.

Modifiers are typically used to enforce access control or validate input parameters, but they can also be used for other purposes such as logging or updating contract state.