Require Function
In Solidity, the require statement is a built-in function that allows you to validate certain conditions before executing a function. If the condition specified in the require statement evaluates to false, the transaction will be reverted, and any changes made during the transaction will be discarded.

In this example, the require statement checks whether num2 is not equal to zero. If num2 is equal to zero, the transaction will be reverted with the error message "Cannot divide by zero". If num2 is not equal to zero, the function will proceed with the division operation and return the result.

The require statement is often used to validate inputs to a function or to check the state of the contract before executing a function. It is an important security feature in Solidity that helps prevent unexpected behavior and protects against certain types of attacks.

Attention: 
Both require and modifiers are used to add conditions to Solidity functions, but they serve slightly different purposes and are not interchangeable.
require statements are used to validate inputs or intermediate values inside a function. If the condition in a require statement is not met, the function will immediately revert and any changes made by the function will be undone.
Modifiers, on the other hand, are used to apply conditions to entire functions or function sets. They are typically used to restrict access to certain functions or to enforce certain pre-conditions on a set of related functions. Modifiers are defined separately from functions and can be applied to multiple functions.