Mapping
A mapping is a key-value data structure similar to a dictionary or hash map in other programming languages. A mapping allows you to associate a value with a key and quickly look up the value by its corresponding key. Mappings are typically used to store and retrieve data efficiently in a smart contract.