Send
The send function is another way to send ether in Solidity, but it's a lower-level function than transfer and has some limitations. 
In this example, the sendEther function sends the amount of ether specified in msg.value to the address recipient using the send function. The send function returns a boolean value indicating whether the transaction was successful, which we check using the require statement.