Call
The call function is a low-level way to invoke a function in another contract and can be used to send ether, transfer tokens, or perform any other type of transaction.
In this example, the callOtherContract function invokes the transfer function in another contract specified by the otherContract address and passes in the current user's address and the amount value as arguments. The call function is used to execute this transaction, and the value field is used to specify the amount of ether to send along with the transaction. The abi.encodeWithSignature function is used to encode the function call with its arguments in a way that can be understood by the receiving contract. Finally, we check the success of the transaction using the require statement.